Add files via upload
We want to predict whether users will cancel their tickets or not. For each cancellation, ''mrbilit.com'' has to pay a penalty to the airline! So canceled tickets mean incurring costs for the company. Therefore, identifying tickets that are likely to be canceled will greatly help the company’s risk management.